refactor working
